+++ src/main/java/com/unicredit/cap/busineslogic/PlacementService.java
@@ -83,7 +83,7 @@ public class PlacementService {
 		   		PlacementTransfer  transfer = new PlacementTransfer();
 		   		transfer.setDateFrom(new Date());
 		   		transfer.setToOrg(placement.getCreatingOrg());
-		   		transfer.setToUser(placement.getCreateUser().intValue());
+		   		transfer.setToUser(placement.getCreateUser().longValue());
 		   		transfer.setMovementType("Kreiranje Plasmana");  
 		   		transfer.setPlacement(placement);
